+++ src/TrolleyBus.java
@@ -1,11 +1,11 @@
 public class TrolleyBus extends Vehicle {
-    private int electricConsumption;
-    TrolleyBus(double width, double length, int weight, int maxWeight, GearboxType gearboxType, Engine.FuelType fuelType, int enginePower, int electricConsumption) {
+    private double electricConsumption;
+    TrolleyBus(double width, double length, int weight, int maxWeight, GearboxType gearboxType, Engine.FuelType fuelType, int enginePower, double electricConsumption) {
         super(width, length, weight, maxWeight, gearboxType, fuelType, enginePower);
         this.electricConsumption = electricConsumption;
     }
 
-    public int getElectricConsumption() {
+    public double getElectricConsumption() {
         return electricConsumption;
     }
 }
